requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
drf-spectacular>=0.27
Pillow>=10.0
django-filter>=24.0
resend
pytest
pytest-django
flake8
isort